Added fail safe checks and refactoring
